chore: prepare VIP-652 for proposal
